+++ src/main/kotlin/frc/team449/control/holonomic/SwerveDrive.kt
@@ -13,7 +13,7 @@
  import edu.wpi.first.math.kinematics.SwerveModuleState
  import edu.wpi.first.wpilibj.Timer
  import edu.wpi.first.wpilibj2.command.SubsystemBase
- import frc.team449.robot2022.drive.DriveConstants
+ import frc.team449.robot2022.RobotConstants
  import frc.team449.robot2022.drive.SwerveConstants
  import frc.team449.robot2022.vision.VisionConstants
  import frc.team449.system.AHRS
@@ -49,7 +49,7 @@
     kinematics,
     ahrs.heading,
     getPositions(),
-    DriveConstants.INITIAL_POSE,
+    RobotConstants.INITIAL_POSE,
     MatBuilder(Nat.N3(), Nat.N1()).fill(.005, .005, .005), // [theta, fl_pos, fr_pos, bl_pos, br_pos]
     MatBuilder(Nat.N3(), Nat.N1()).fill(.05, .075, .025) // [x, y, theta]
   )
@@ -228,8 +228,8 @@
        return SwerveDrive(
          modules,
          ahrs,
-         DriveConstants.MAX_LINEAR_SPEED,
-         DriveConstants.MAX_ROT_SPEED,
+         RobotConstants.MAX_LINEAR_SPEED,
+         RobotConstants.MAX_ROT_SPEED,
          VisionConstants.CAMERAS
        )
      }
